+++ .gitignore
@@ -11,3 +11,4 @@ init-js-tool/.env
 .env
 node_modules
 /contract-records.json
+yarn-error.log
