+++ init-js-tool/types.ts
@@ -1,0 +1,3 @@
+import { ClarityType } from '@stacks/transactions';
+
+export type ClarityResult = { data: { [key: string]: any }; type: ClarityType };
