+++ init-js-tool/wallet.ts
@@ -1,26 +1,19 @@
-const { generateWallet } = require('@stacks/wallet-sdk');
-const {
-  makeSTXTokenTransfer,
-  broadcastTransaction,
-} = require('@stacks/transactions');
-const { wait_until_confirmation } = require('./utils');
-const BigNum = require('bn.js');
-const {
-  STACKS_API_URL,
+import { broadcastTransaction, makeSTXTokenTransfer } from '@stacks/transactions';
+import BigNum from 'bn.js';
+import { StacksMocknet } from '@stacks/network/dist';
+import { generateWallet } from '@stacks/wallet-sdk/dist';
+
+import {
   DEPLOYER_ACCOUNT_ADDRESS,
-  DEPLOYER_ACCOUNT_SECRET,
   DEPLOYER_ACCOUNT_PASSWORD,
-  USER_ACCOUNT_SECRET,
+  DEPLOYER_ACCOUNT_SECRET,
+  STACKS_API_URL,
   USER_ACCOUNT_PASSWORD,
-} = require('./constants');
+  USER_ACCOUNT_SECRET,
+} from './constants';
+import { wait_until_confirmation } from './utils';
 
-const {
-  StacksRegtest,
-  StacksMocknet,
-  StacksMainnet,
-} = require('@stacks/network');
-
-const getDeployerPK = async () => {
+export const getDeployerPK = async () => {
   const wallet = await generateWallet({
     secretKey: DEPLOYER_ACCOUNT_SECRET(),
     password: DEPLOYER_ACCOUNT_PASSWORD(),
@@ -29,7 +22,7 @@ const getDeployerPK = async () => {
   return privateKey;
 };
 
-const getUserPK = async () => {
+export const getUserPK = async () => {
   const wallet = await generateWallet({
     secretKey: USER_ACCOUNT_SECRET(),
     password: USER_ACCOUNT_PASSWORD(),
@@ -39,12 +32,12 @@ const getUserPK = async () => {
 };
 
 // const network = new StacksRegtest({
-const network = new StacksMocknet({
+export const network = new StacksMocknet({
   // const network = new StacksMainnet({
   url: STACKS_API_URL(),
 });
 
-const genesis_transfer = async () => {
+export const genesis_transfer = async () => {
   console.log('genesis-transfer...');
 
   const txOptions = {
@@ -53,7 +46,8 @@ const genesis_transfer = async () => {
     senderKey:
       '21d43d2ae0da1d9d04cfcaac7d397a33733881081f0b2cd038062cf0ccbb752601',
     network: network,
-  };
+  } as any;
+  // FIXME: typescript migrate;
 
   try {
     const transaction = await makeSTXTokenTransfer(txOptions);
@@ -64,8 +58,3 @@ const genesis_transfer = async () => {
     console.log(error);
   }
 };
-
-exports.getDeployerPK = getDeployerPK;
-exports.getUserPK = getUserPK;
-exports.network = network;
-exports.genesis_transfer = genesis_transfer;
